aucto-main/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart
pyyaml
httpx
pandas
google-generativeai
scikit-learn
joblib
numpy



